requirements.txt
openai
python-dotenv
requests
simple-term-menu
Pillow
fade